fix: prevent repeat initial storage loads
