+++ packages/core/src/contexts/auth.tsx
@@ -24,6 +24,7 @@ import { hashPIN } from '../utils/crypto'
 import { migrateToAskar } from '../utils/migration'
 
 export interface AuthContext {
+  lockOutUser: () => void
   checkWalletPIN: (PIN: string) => Promise<boolean>
   getWalletSecret: () => Promise<WalletSecret | undefined>
   walletSecret?: WalletSecret
@@ -128,6 +129,18 @@ export const AuthProvider: React.FC<React.PropsWithChildren> = ({ children }) =>
     setWalletSecret(undefined)
   }, [])
 
+  const lockOutUser = useCallback(() => {
+    removeSavedWalletSecret()
+    dispatch({
+      type: DispatchAction.DID_AUTHENTICATE,
+      payload: [false],
+    })
+    dispatch({
+      type: DispatchAction.LOCKOUT_UPDATED,
+      payload: [{ displayNotification: true }],
+    })
+  }, [removeSavedWalletSecret, dispatch])
+
   const disableBiometrics = useCallback(async () => {
     await wipeWalletKey(true)
   }, [])
@@ -169,6 +182,7 @@ export const AuthProvider: React.FC<React.PropsWithChildren> = ({ children }) =>
   return (
     <AuthContext.Provider
       value={{
+        lockOutUser,
         checkWalletPIN,
         getWalletSecret,
         removeSavedWalletSecret,
